:lock: file Remove and rename

+++ sockets/removeSpecListener.js
@@ -2,12 +2,16 @@ const getPathToSpec = require('../utils/node/getPathToSpec');
 const util = require('util');
 const fs = require('fs');
 const path = require('path');
+const filenameRegex = require('../utils/regex').filename;
 
 const listenerName = 'removeSpec';
 
 const unlink = util.promisify(fs.unlink);
 
 async function removeSpecListener(nodeId, filename, session) {
+  console.log(filename);
+  if (!filenameRegex.test(filename)) throw Error('Invalid filename');
+
   const p = await getPathToSpec(nodeId, session.currentWorkspace);
   const ret = await unlink(path.join(p, filename));
   return !(ret);
